+++ sdl2/ext/resources.py
@@ -20,16 +20,24 @@ else:
 
 
 def open_zipfile(archive, filename, directory=None):
-    """Opens and reads a certain file from a ZIP archive.
+    """Retrieves a given file from a ZIP archive.
 
-    Opens and reads a certain file from a ZIP archive. The result is
-    returned as StringIO stream. filename can be a relative or absolute
-    path within the ZIP archive. The optional directory argument can be
-    used to supply a relative directory path, under which filename will
-    be searched.
+    Args:
+        archive (:obj:`~zipfile.ZipFile`, str): The ZipFile object or path to
+            the ZIP archive containing the desired file.
+        filename (str): The name of the file to retrieve from the archive.
+        directory (str, optional): The path to the directory within the archive
+            containing the file to retrieve. Defaults to the root level of the
+            archive.
 
-    If the filename could not be found, a KeyError will be raised.
-    Raises a TypeError, if archive is not a valid ZIP archive.
+    Returns:
+        :obj:`~io.BytesIO`: A Python bytestream object containing the requested
+        file.
+
+    Raises:
+        KeyError: If the given file could not be found within the archive.
+        TypeError: If the archive is not a valid ZIP archive.
+
     """
     data = None
     opened = False
@@ -55,42 +63,46 @@ def open_zipfile(archive, filename, directory=None):
 
 
 def open_tarfile(archive, filename, directory=None, ftype=None):
-    """Opens and reads a certain file from a TAR archive.
+    """Retrieves a given file from a TAR archive.
 
-    Opens and reads a certain file from a TAR archive. The result is
-    returned as StringIO stream. filename can be a relative or absolute
-    path within the TAR archive. The optional directory argument can be
-    used to supply a relative directory path, under which filename will
-    be searched.
+    If the TAR archive uses ``.tar.gz`` or ``.tar.bz2`` compression and the
+    file name does not contain either of these identifiers, the compression
+    type must be manually specified.
 
-    ftype is used to supply additional compression information, in case
-    the system cannot determine the compression type itself, and can be
-    either 'gz' for gzip compression or 'bz2' for bzip2 compression.
+    Args:
+        archive (:obj:`~tarfile.TarFile`, str): The TarFile object or path to
+            the TAR archive containing the desired file.
+        filename (str): The name of the file to retrieve from the archive.
+        directory (str, optional): The path to the directory within the archive
+            containing the file to retrieve. Defaults to the root level of the
+            archive.
+        ftype (str, optional): The compression type (if any) used for the TAR
+            file, can be either 'gz', 'bz2', or None (no compression). If not
+            specified, will default to assuming no compression.
 
-    Note:
+    Returns:
+        :obj:`~io.BytesIO`: A Python bytestream object containing the requested
+        file.
 
-      If ftype is supplied, the compression mode will be enforced for
-      opening and reading.
+    Raises:
+        KeyError: If the given file could not be found within the archive.
+        TypeError: If the archive is not a supported TAR archive.
 
-    If the filename could not be found or an error occured on reading it,
-    None will be returned.
-
-    Raises a TypeError, if archive is not a valid TAR archive or if type
-    is not a valid value of ('gz', 'bz2').
     """
     data = None
     opened = False
-
-    mode = 'r'
-    if ftype:
-        if ftype not in ('gz', 'bz2'):
-            raise TypeError("invalid TAR compression type")
-        mode = "r:%s" % ftype
 
     if not isinstance(archive, tarfile.TarFile):
         if not tarfile.is_tarfile(archive):
             raise TypeError("passed file does not seem to be a TAR archive")
         else:
+            file_ext = archive.split('.')[-1]
+            if not ftype and file_ext in ('gz', 'bz2'):
+                ftype = file_ext
+            if ftype and ftype not in ('gz', 'bz2'):
+                e = "invalid TAR compression type '{0}' (must be 'gz' or 'bz2')"
+                raise TypeError(e.format(ftype))
+            mode = 'r:{0}'.format(ftype) if ftype else 'r'
             archive = tarfile.open(archive, mode)
             opened = True
 
@@ -108,14 +120,9 @@ def open_tarfile(archive, filename, directory=None, ftype=None):
 
 
 def open_url(filename, basepath=None):
-    """Opens and reads a certain file from a web or remote location.
-
-    Opens and reads a certain file from a web or remote location. This
-    function utilizes the urllib2 module, which means that it is
-    restricted to the types of remote locations supported by urllib2.
-
-    basepath can be used to supply an additional location prefix.
-    """
+    # Opens and reads a certain file from a web or remote location.
+    # Deprecated because its argument names are confusing and because
+    # users are probably better off using urllib directly.
     url = filename
     if basepath:
         url = urlparse.urljoin(basepath, filename)
@@ -123,17 +130,21 @@ def open_url(filename, basepath=None):
 
 
 class Resources(object):
-    """The Resources class manages a set of file resources and eases
-    accessing them by using relative paths, scanning archives
-    automatically and so on.
+    """A container class for managing application resource files.
+    
+    This class eases access to resources by allowing access using relative
+    paths, scanning archives to locate files, and more.
+
+    Args:
+        path (str, optional): The path of a resource directory with which to
+            initialze the container. Defaults to ``None``.
+        subdir (str, optional): Deprecated, do not use.
+        excludepattern (str, optional): A regular expression indicating
+            which directories (if any) to ignore if initializing the
+            container with a resource path. Defaults to ``None``.
+
     """
     def __init__(self, path=None, subdir=None, excludepattern=None):
-        """Creates a new resource container instance.
-
-        If path is provided, the resource container will scan the path
-        and add all found files to itself by invoking
-        scan(path, subdir, excludepattern).
-        """
         self.files = {}
         if path:
             self.scan(path, subdir, excludepattern)
@@ -158,11 +169,13 @@ class Resources(object):
         """
         if not tarfile.is_tarfile(filename):
             raise TypeError("file '%s' is not a valid TAR archive" % filename)
-        mode = 'r'
-        if ftype:
-            if ftype not in ('gz', 'bz2'):
-                raise TypeError("invalid TAR compression type")
-            mode = "r:%s" % ftype
+        file_ext = filename.split('.')[-1]
+        if not ftype and file_ext in ('gz', 'bz2'):
+            ftype = file_ext
+        if ftype and ftype not in ('gz', 'bz2'):
+            e = "invalid TAR compression type '{0}' (must be 'gz' or 'bz2')"
+            raise TypeError(e.format(ftype))
+        mode = 'r:{0}'.format(ftype) if ftype else 'r'
         archname = os.path.abspath(filename)
         archtype = 'tar'
         if ftype:
@@ -176,9 +189,16 @@ class Resources(object):
     def add(self, filename):
         """Adds a file to the Resources container.
 
-        Depending on the file type (determined by the file suffix or name),
-        the file will be automatically scanned (if it is an archive) or
-        checked for availability (if it is a stream/network resource).
+        If the given file is a supported archive, its contents will be scanned
+        and added to the container.
+
+        Args:
+            filename (str): The filepath of the resource to add to the
+                container.
+
+        Raises:
+            ValueError: If the file does not exist at the provided path.
+
         """
         if not os.path.exists(filename):
             raise ValueError("invalid file path")
@@ -190,10 +210,18 @@ class Resources(object):
             self.add_file(filename)
 
     def add_file(self, filename):
-        """Adds a file to the Resources container.
+        """Adds a file without scanning to the Resources container.
 
-        This will only add the passed file and do not scan an archive or
-        check a stream for availability.
+        Unlike :meth:`add`, this method will not attempt to add the contents
+        of any provided archives to the container.
+
+        Args:
+            filename (str): The filepath of the resource to add to the
+                container.
+
+        Raises:
+            ValueError: If the file does not exist at the provided path.
+
         """
         if not os.path.exists(filename):
             raise ValueError("invalid file path")
@@ -204,28 +232,57 @@ class Resources(object):
         self.files[fname] = (None, None, abspath)
 
     def add_archive(self, filename, typehint='zip'):
-        """Adds an archive file to the Resources container.
+        """Adds a ``.zip`` or ``.tar`` archive to the container.
 
         This will scan the passed archive and add its contents to the
-        list of available resources.
+        list of available resources. Currently ``.zip``, ``.tar``,
+        ``.tar.bz2``, and ``.tar.gz`` formats are supported.
+
+        Args:
+            filename (str): The filepath of the archive to scan and add to the
+                container.
+            typehint (str, optional): The format of the archive to add to the
+                container, required if using a custom file extension. Must be
+                one of ``zip``, ``tar``, ``tarbz2``, or ``targz``. Defaults to
+                ``zip`` if not specified.
+
+        Raises:
+            ValueError: If the file does not exist at the provided path, or if
+                the file is not a supported archive type.
+
         """
         if not os.path.exists(filename):
             raise ValueError("invalid file path")
-        if typehint == 'zip':
+        fname = os.path.basename(filename)
+        if 'zip' in fname.split('.'):
             self._scanzip(filename)
-        elif typehint == 'tar':
+        elif 'tar' in fname.split('.'):
             self._scantar(filename)
-        elif typehint == 'tarbz2':
-            self._scantar(filename, 'bz2')
-        elif typehint == 'targz':
-            self._scantar(filename, 'gz')
         else:
-            raise ValueError("unsupported archive type")
+            if typehint == 'zip':
+                self._scanzip(filename)
+            elif typehint == 'tar':
+                self._scantar(filename)
+            elif typehint == 'tarbz2':
+                self._scantar(filename, 'bz2')
+            elif typehint == 'targz':
+                self._scantar(filename, 'gz')
+            else:
+                raise ValueError("unsupported archive type")
 
     def get(self, filename):
-        """Gets a specific file from the Resources.
+        """Retrieves a resource file by name from the container.
 
-        Raises a KeyError, if filename could not be found.
+        Args:
+            filename (str): The file name of the resource to retrieve.
+
+        Returns:
+            :obj:`~io.BytesIO`: A Python bytestream object containing the
+            retrieved resource file.
+
+        Raises:
+            KeyError: If the given file could not be found.
+
         """
         archive, ftype, pathname = self.files[filename]
         if archive:
@@ -245,14 +302,7 @@ class Resources(object):
         return data
 
     def get_filelike(self, filename):
-        """Like get(), but tries to return the original file handle, if
-        possible.
-
-        If the passed filename is only available within an archive, a
-        StringIO instance will be returned.
-
-        Raises a KeyError, if filename could not be found.
-        """
+        # Deprecated, doesn't make much difference in Python 3
         archive, ftype, pathname = self.files[filename]
         if archive:
             if ftype == 'zip':
@@ -268,12 +318,21 @@ class Resources(object):
         return open(pathname, 'rb')
 
     def get_path(self, filename):
-        """Gets the path of the passed filename.
+        """Gets the path of a given resource file.
 
-        If filename is only available within an archive, a string in
-        the form 'filename@archivename' will be returned.
+        If the file is only available within an archive, a string in the form
+        ``filename@archivename`` will be returned.
 
-        Raises a KeyError, if filename could not be found.
+        Args:
+            filename (str): The file name of the resource to locate.
+
+        Returns:
+            str: The absolute path of the resource file, or the archive
+            identifier string if the resource is inside an archive.
+
+        Raises:
+            KeyError: If the given file could not be found.
+
         """
         archive, ftype, pathname = self.files[filename]
         if archive:
@@ -281,21 +340,21 @@ class Resources(object):
         return pathname
 
     def scan(self, path, subdir=None, excludepattern=None):
-        """Scans a path and adds all found files to the Resources
-        container.
+        """Scans a path, adding all matching files to the container.
 
-        Scans a path and adds all found files to the Resources
-        container. If a file is a supported (ZIP or TAR) archive, its
-        contents will be indexed and added automatically.
+        If a located file is a ``.zip`` or ``.tar`` archive, its
+        contents will be indexed and added to the container automatically.
 
-        The method will consider the directory part (os.path.dirname) of
-        the provided path as path to scan, if the path is not a
-        directory. If subdir is provided, it will be appended to the
-        path and used as starting point for adding files to the
-        Resources container.
+        Args:
+            path (str): The path of the directory to scan.
+            subdir (str, optional): Deprecated, do not use.
+            excludepattern (str, optional): A regular expression indicating
+                which directories (if any) within the file structure of the
+                given path to exclude from indexing. Defaults to ``None``. 
 
-        excludepattern can be a regular expression to skip directories, which
-        match the pattern.
+        Raises:
+            ValueError: If the specified path does not exist.
+
         """
         match = None
         if excludepattern: